adjustments to two tests
The previous CIDs being matched in the test do not exist in the PubChem database. Consequently, when testing on the PubChem website directly instead of using the BrokenRecord files, the tests were failing. With this change tests do not fail anymore, as I replaced the erroneous CID with one that exists in PubChem.